implemented userrole with token bu jwt and checking accesss to important settings

+++ src/backend/VoltStream.Application/Features/Categories/Commands/DeleteCategoryCommand.cs
@@ -8,25 +8,24 @@ using VoltStream.Domain.Entities;
 
 public record DeleteCategoryCommand(long Id) : IRequest<bool>;
 
-public class DeleteCategoryCommandHandler(
-    IAppDbContext context)
+public class DeleteCategoryCommandHandler(IAppDbContext context)
     : IRequestHandler<DeleteCategoryCommand, bool>
 {
     public async Task<bool> Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
     {
-        var category = await context.Categories.FirstOrDefaultAsync(p => p.Id == request.Id, cancellationToken)
+        var category = await context.Categories
+            .FirstOrDefaultAsync(c => c.Id == request.Id, cancellationToken)
             ?? throw new NotFoundException(nameof(Category), nameof(request.Id), request.Id);
 
-        context.Categories.Remove(category);
+        var hasSoldProducts = await context.Products
+            .Where(p => p.CategoryId == request.Id)
+            .AnyAsync(p => context.SaleItems.Any(si => si.ProductId == p.Id), cancellationToken);
 
-        try
-        {
-            await context.SaveAsync(cancellationToken);
-        }
-        catch (DbUpdateException)
-        {
-            throw new ForbiddenException("Ushbu kategoriyada mahsulotlar mavjud yoki tizim xatoligi tufayli o'chirib bo'lmaydi.");
-        }
+        if (hasSoldProducts)
+            throw new ForbiddenException("Kategoriyani o'chirib bo'lmaydi: Undagi ayrim mahsulotlar savdoda ishtirok etgan.");
+
+        context.Categories.Remove(category);
+        await context.SaveAsync(cancellationToken);
 
         return true;
     }
